requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.7.0
aiofiles==23.2.1

# AI / RAG
groq==0.9.0
pinecone-client==3.2.2
sentence-transformers==2.7.0

langchain==0.2.0
langchain-groq==0.1.6
langchain-pinecone==0.1.2
dotenv